+++ src/tables/mod.rs
@@ -38,6 +38,7 @@ pub mod loca;
 pub mod ltsh;
 pub mod math;
 pub mod maxp;
+pub mod merg;
 pub mod meta;
 pub mod mvar;
 pub mod name;